+++ packages/core-data/src/resolvers.js
@@ -15,7 +15,13 @@ import apiFetch from '@wordpress/api-fetch';
  */
 import { STORE_NAME } from './name';
 import { getOrLoadEntitiesConfig, DEFAULT_ENTITY_KEY } from './entities';
-import { forwardResolver, getNormalizedCommaSeparable } from './utils';
+import {
+	forwardResolver,
+	getNormalizedCommaSeparable,
+	getUserPermissionCacheKey,
+	getUserPermissionsFromResponse,
+	ALLOWED_RESOURCE_ACTIONS,
+} from './utils';
 import { getSyncProvider } from './sync';
 import { fetchBlockPatterns } from './fetch';
 
@@ -167,35 +173,16 @@ export const getEntityRecord =
 
 				const response = await apiFetch( { path, parse: false } );
 				const record = await response.json();
-
-				const allowHeader = response.headers?.get( 'allow' );
-				const allowedMethods = allowHeader?.allow || allowHeader || '';
-				const permissions = {};
-				const methods = {
-					create: 'POST',
-					read: 'GET',
-					update: 'PUT',
-					delete: 'DELETE',
-				};
-				for ( const [ actionName, methodName ] of Object.entries(
-					methods
-				) ) {
-					permissions[ actionName ] =
-						allowedMethods.includes( methodName );
-				}
+				const permissions = getUserPermissionsFromResponse( response );
 
 				registry.batch( () => {
 					dispatch.receiveEntityRecords( kind, name, record, query );
 
-					for ( const action of [
-						'create',
-						'read',
-						'update',
-						'delete',
-					] ) {
-						const permissionKey = [ action, kind, name, key ]
-							.filter( Boolean )
-							.join( '/' );
+					for ( const action of ALLOWED_RESOURCE_ACTIONS ) {
+						const permissionKey = getUserPermissionCacheKey(
+							action,
+							{ kind, name, id: key }
+						);
 
 						dispatch.receiveUserPermission(
 							permissionKey,
@@ -395,9 +382,7 @@ export const getEmbedPreview =
 export const canUser =
 	( requestedAction, resource, id ) =>
 	async ( { dispatch, registry } ) => {
-		const retrievedActions = [ 'create', 'read', 'update', 'delete' ];
-
-		if ( ! retrievedActions.includes( requestedAction ) ) {
+		if ( ! ALLOWED_RESOURCE_ACTIONS.includes( requestedAction ) ) {
 			throw new Error( `'${ requestedAction }' is not a valid action.` );
 		}
 
@@ -429,7 +414,7 @@ export const canUser =
 		const { hasStartedResolution } = registry.select( STORE_NAME );
 
 		// Prevent resolving the same resource twice.
-		for ( const relatedAction of retrievedActions ) {
+		for ( const relatedAction of ALLOWED_RESOURCE_ACTIONS ) {
 			if ( relatedAction === requestedAction ) {
 				continue;
 			}
@@ -456,31 +441,10 @@ export const canUser =
 			return;
 		}
 
-		// Optional chaining operator is used here because the API requests don't
-		// return the expected result in the native version. Instead, API requests
-		// only return the result, without including response properties like the headers.
-		const allowedMethods = response.headers?.get( 'allow' ) || '';
-
-		const permissions = {};
-		const methods = {
-			create: 'POST',
-			read: 'GET',
-			update: 'PUT',
-			delete: 'DELETE',
-		};
-		for ( const [ actionName, methodName ] of Object.entries( methods ) ) {
-			permissions[ actionName ] = allowedMethods.includes( methodName );
-		}
-
+		const permissions = getUserPermissionsFromResponse( response );
 		registry.batch( () => {
-			for ( const action of retrievedActions ) {
-				const key = (
-					typeof resource === 'object'
-						? [ action, resource.kind, resource.name, resource.id ]
-						: [ action, resource, id ]
-				)
-					.filter( Boolean )
-					.join( '/' );
+			for ( const action of ALLOWED_RESOURCE_ACTIONS ) {
+				const key = getUserPermissionCacheKey( action, resource, id );
 
 				dispatch.receiveUserPermission( key, permissions[ action ] );
